[models.py and views.py]Touch up price decimal in models. Touch up css and grab info from db in profile page

+++ main_app/models.py
@@ -59,9 +59,8 @@ class Order(models.Model):
 class OrderItem(models.Model):
     order = models.ForeignKey(Order, related_name='items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, related_name='order_items', on_delete=models.CASCADE)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
     quantity = models.IntegerField()
-    # amount = models.IntegerField()
 
     def __str__(self):
         return '{}'.format(self.id)
